requirements.txt
fastapi==0.128.6
uvicorn[standard]==0.34.0
pydantic==2.8.2
pydantic-settings==2.12.0
python-dotenv==1.0.1
python-multipart==0.0.20

SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

requests==2.32.5
httpx==0.28.1
orjson==3.10.16
ujson==5.10.0

numpy==1.26.4
pandas==2.2.3
faiss-cpu==1.13.2
rank-bm25==0.2.2

sentence-transformers==5.2.3
transformers==4.41.2
torch==2.1.0
huggingface-hub==0.23.4
tiktoken==0.12.0

openai==2.21.0

langchain==1.2.9
langchain-core==1.2.13
langchain-community==0.4.1
langchain-openai==1.1.10
langgraph==1.0.8